+++ CHANGELOG.md
@@ -7,6 +7,7 @@ ChangeLog
 
 - IMPROVEMENT: allow one to use the REST API of the server to control the UI
 - IMPROVEMENT: allow one to use keystrokes to control the UI
+- IMPROVEMENT: send HeyGen a pseudo task regularly to keep connection alive
 - BUGFIX: fix startup scripts
 - BUGFIX: fix text-area scrolling under Windows
 - BUGFIX: avoid outlines for sections/tabs